Events may emit incorrect output amounts
